blog-posts/databricks-officeqa-benchmark/pyproject.toml
[project]
name = "databricks-officeqa-benchmark"
version = "0.1.0"
description = "Databricks OfficeQA benchmark experiments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.56.0",
    "google-cloud-aiplatform>=1.113.0",
    "modal>=1.3.0.post1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
